refactoring generation of descriptions, create dobBlockGenerator
